Removed generic files and gave them designated files. e.g: utils.py and docker-compose-general
